Div Changes & ParallaxFloat fix
SpinePool Load. SpriteDraw w Offset. Stuff.

+++ DragonEngine/Entities/Parallax/ParallaxPlane.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace DragonEngine.Entities
 {
@@ -49,8 +50,8 @@ namespace DragonEngine.Entities
         /// <param name="pCamera">Zu verwendende Kamera.</param>
         public virtual void Update(Camera pCamera)
         {
-            PositionX = pCamera.PositionX - ((Width - pCamera.Width) * ((pCamera.PositionX - pCamera.ViewArea.X) / (pCamera.ViewArea.Width - pCamera.Width)));
-            PositionY = pCamera.PositionY - ((Height - pCamera.Height) * ((pCamera.PositionY - pCamera.ViewArea.Y) / (pCamera.ViewArea.Height - pCamera.Height)));
+            PositionX = pCamera.PositionX - (int)((float)(Width - pCamera.Width) * ((float)(pCamera.PositionX - pCamera.ViewArea.X) / (float)(pCamera.ViewArea.Width - pCamera.Width)));
+            PositionY = pCamera.PositionY - (int)((float)(Height - pCamera.Height) * ((float)(pCamera.PositionY - pCamera.ViewArea.Y) / (float)(pCamera.ViewArea.Height - pCamera.Height)));
         }
 
         #endregion
